refactor(l5): move promotion-nomination and sharing from src/ into server/domain (issue 864)
Two modules move out of src/ into the server/domain layer they already
belong to, each leaving a re-export shim at the old path so callers
outside this change keep resolving:

- src/promotion-nomination.ts -> server/domain/promotion-nomination.ts
- src/sharing.ts              -> server/domain/sharing.ts
  (src/sharing.test.ts        -> server/domain/sharing.test.ts)

server/tools/promotion.ts, server/tools/scan-namespace.ts, and
server/tools/ingest-conversation-facts.ts import the new paths directly
rather than going through a shim.

sharing.ts needed structural work to pass the server-side lint gate, so
classifyShareCandidate and shareRejectionDetail were relieved of inline
work by hoisting named helpers: detectSecret, detectPrivate,
buildShareRejectionDetail, shareDecisionLogContext, detectorLogFields,
and noiseRejectionReason. Behavior is unchanged and the moved test
covers it at the same assertions.

Two modules were considered and deliberately deferred:

- tiering.ts stays in src/. Bringing it to standard needs two
  max-params fixes, and those pull scripts/tier-lane-durable.ts and
  scripts/done-means/605-chunk-write-routing.driver.ts (28 pre-existing
  findings between them) into the gate. Those two scripts get brought to
  standard first, in their own change.
- promotion-service.ts stays in src/. It reads
  OPENBRAIN_PROMOTION_KILL_SWITCH from process.env directly, which the
  server-side rule forbids; routing that through config is its own
  change, not a side effect of a move.

The two shims are excluded from the closure count, which moves 67 -> 65.

Issue: #864
